2025/BackdoorCTF/MCP Mayhem/challenge/web/requirements.txt
Flask
pyjwt
mcp
requests